Ensure event mapper is not a virtual connector and tidy up audit log messages
