+++ open-metadata-implementation/adapters/open-connectors/connector-configuration-factory/src/main/java/org/odpi/openmetadata/adapters/repositoryservices/ConnectorConfigurationFactory.java
@@ -659,7 +659,7 @@ public class ConnectorConfigurationFactory
         final String connectorTypeDescription   = "OMRS default enterprise connector type.";
         final String connectorTypeJavaClassName = OMRSTopicProvider.class.getName();
 
-        String connectorTypeName = "DefaultEnterpriseTopic.ConnectorType." + localServerName;
+        String connectorTypeName = "EnterpriseTopicConnectorType.Server." + localServerName;
 
         ConnectorType connectorType = new ConnectorType();
 
@@ -673,7 +673,7 @@ public class ConnectorConfigurationFactory
 
         final String connectionDescription = "OMRS default enterprise topic connection.";
 
-        String connectionName = "DefaultEnterpriseTopic.Connection." + localServerName;
+        String connectionName = "EnterpriseTopicConnector.Server." + localServerName;
 
         VirtualConnection connection = new VirtualConnection();
 
@@ -730,7 +730,7 @@ public class ConnectorConfigurationFactory
 
         final String connectionDescription = "OMRS default cohort topic connection.";
 
-        String connectionName = "DefaultCohortTopic.Connection." + cohortName;
+        String connectionName = "TopicConnector.Cohort." + cohortName;
 
         VirtualConnection connection = new VirtualConnection();
