+++ open-metadata-implementation/adapters/open-connectors/repository-services-connectors/open-metadata-collection-store-connectors/graph-repository-connector/src/main/java/org/odpi/openmetadata/adapters/repositoryservices/graphrepository/eventmapper/GraphOMRSRepositoryEventMapper.java
@@ -1,16 +1,29 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 package org.odpi.openmetadata.adapters.repositoryservices.graphrepository.eventmapper;
 
+import org.odpi.openmetadata.frameworks.connectors.Connector;
+import org.odpi.openmetadata.frameworks.connectors.VirtualConnectorExtension;
 import org.odpi.openmetadata.frameworks.connectors.ffdc.ConnectorCheckedException;
+import org.odpi.openmetadata.repositoryservices.auditlog.OMRSAuditCode;
+import org.odpi.openmetadata.repositoryservices.auditlog.OMRSAuditLog;
+import org.odpi.openmetadata.repositoryservices.auditlog.OMRSAuditingComponent;
+import org.odpi.openmetadata.repositoryservices.connectors.openmetadatatopic.OpenMetadataTopicConnector;
 import org.odpi.openmetadata.repositoryservices.connectors.stores.metadatacollectionstore.repositoryeventmapper.OMRSRepositoryEventMapperBase;
+
+import java.util.ArrayList;
+import java.util.List;
 
 
 /**
  * GraphOMRSRepositoryEventMapper supports the event mapper function for the default graph store open
  * metadata repository.
  */
-public class GraphOMRSRepositoryEventMapper extends OMRSRepositoryEventMapperBase
+public class GraphOMRSRepositoryEventMapper extends OMRSRepositoryEventMapperBase implements VirtualConnectorExtension
 {
+    private static final OMRSAuditLog auditLog = new OMRSAuditLog(OMRSAuditingComponent.LOCAL_REPOSITORY_EVENT_MAPPER);
+
+    private List<OpenMetadataTopicConnector> eventBusConnectors = new ArrayList<>();
+
     /**
      * Default constructor
      */
@@ -29,6 +42,63 @@ public class GraphOMRSRepositoryEventMapper extends OMRSRepositoryEventMapperBas
     {
         super.start();
     }
+
+    /**
+     * Registers itself as a listener of any OpenMetadataTopicConnectors that are passed as
+     * embedded connectors.
+     *
+     * @param embeddedConnectors  list of connectors
+     */
+    public void initializeEmbeddedConnectors(List<Connector> embeddedConnectors)
+    {
+        final String  methodName = "initializeEmbeddedConnectors";
+
+        /*
+         * Step through the embedded connectors, selecting only the OpenMetadataTopicConnectors
+         * to use.
+         */
+        if (embeddedConnectors != null)
+        {
+            for (Connector  embeddedConnector : embeddedConnectors)
+            {
+                if ((embeddedConnector != null) && (embeddedConnector instanceof OpenMetadataTopicConnector))
+                {
+                    /*
+                     * Successfully found an event bus connector of the right type.
+                     */
+                    OpenMetadataTopicConnector realTopicConnector = (OpenMetadataTopicConnector)embeddedConnector;
+
+                    String   topicName = realTopicConnector.registerListener(this);
+                    this.eventBusConnectors.add(realTopicConnector);
+
+                    OMRSAuditCode auditCode = OMRSAuditCode.EVENT_MAPPER_LISTENER_REGISTERED;
+                    auditLog.logRecord(methodName,
+                                       auditCode.getLogMessageId(),
+                                       auditCode.getSeverity(),
+                                       auditCode.getFormattedLogMessage(repositoryEventMapperName, topicName),
+                                       this.getConnection().toString(),
+                                       auditCode.getSystemAction(),
+                                       auditCode.getUserAction());
+                }
+            }
+        }
+
+        /*
+         * OMRSTopicConnector needs at least one event bus connector to operate successfully.
+         */
+        if (this.eventBusConnectors.isEmpty())
+        {
+            OMRSAuditCode auditCode = OMRSAuditCode.EVENT_MAPPER_LISTENER_DEAF;
+            auditLog.logRecord(methodName,
+                               auditCode.getLogMessageId(),
+                               auditCode.getSeverity(),
+                               auditCode.getFormattedLogMessage(repositoryEventMapperName),
+                               this.getConnection().toString(),
+                               auditCode.getSystemAction(),
+                               auditCode.getUserAction());
+        }
+    }
+
 
     /**
      * Method to pass an event received on topic.
