+++ open-metadata-implementation/repository-services/repository-services-apis/src/main/java/org/odpi/openmetadata/repositoryservices/connectors/openmetadatatopic/OpenMetadataTopicConnector.java
@@ -163,6 +163,17 @@ public abstract class OpenMetadataTopicConnector extends ConnectorBase implement
         if (topicListener != null)
         {
             topicListeners.add(topicListener);
+
+            if (super.connectionProperties != null)
+            {
+                EndpointProperties endpoint = super.connectionProperties.getEndpoint();
+
+                if (endpoint != null)
+                {
+                    topicName = endpoint.getAddress();
+                }
+            }
+
             return topicName;
         }
